fix:delete post to delete related files and comment, and allow to owner to delete and update post

+++ src/main/java/ac/kr/smu/lookie/socialworker/controller/PostController.java
@@ -1,8 +1,9 @@
 package ac.kr.smu.lookie.socialworker.controller;
 
-import ac.kr.smu.lookie.socialworker.domain.Board;
+
 import ac.kr.smu.lookie.socialworker.domain.Post;
 import ac.kr.smu.lookie.socialworker.domain.User;
+import ac.kr.smu.lookie.socialworker.service.CommentService;
 import ac.kr.smu.lookie.socialworker.service.PostService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,7 +18,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Map;
 
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
@@ -31,10 +31,10 @@ public class PostController {
     private final int PAGE_SIZE = 25;
 
     @GetMapping
-    public ResponseEntity<?> getPostList(@PathVariable("boardId") Long boardId,@RequestParam int page){
+    public ResponseEntity<?> getPostList(@PathVariable("boardId") Long boardId,@RequestParam int page){//게시판 글 목록
         Pageable pageable = PageRequest.of(page-1,PAGE_SIZE);
-        Board board = Board.builder().id(boardId).build();
-        Page<Post> postList = postService.getPostList(board, pageable);
+
+        Page<Post> postList = postService.getPostList(boardId, pageable);
         PageMetadata pageMetadata = new PageMetadata(postList.getSize(), postList.getNumber(), postList.getTotalElements());
         PagedModel<Post> body = PagedModel.of(postList.getContent(),pageMetadata);
 
@@ -43,15 +43,16 @@ public class PostController {
     }
 
     @GetMapping("/{postId}")
-    public ResponseEntity<?> getPost(@PathVariable("postId") Long postId){
-        EntityModel<Post> body = EntityModel.of(postService.getPost(postId));
+    public ResponseEntity<?> getPost(@PathVariable("postId") Long postId){//글 상세보기
+        Post post = postService.getPost(postId);
+        EntityModel<Post> body = EntityModel.of(post);
 
         body.add(linkTo(methodOn(PostController.class).getPost(postId)).withSelfRel());
         return ResponseEntity.ok(body);
     }
 
     @PostMapping
-    public ResponseEntity<?> postPost(@PathVariable("boardId") Long boardId, @RequestBody Post post){
+    public ResponseEntity<?> postPost(@PathVariable("boardId") Long boardId, @RequestBody Post post){//글 등록
         EntityModel<Post> body = EntityModel.of(postService.register(post));
 
         body.add(linkTo(methodOn(PostController.class).postPost(boardId,post)).withSelfRel());
@@ -59,10 +60,15 @@ public class PostController {
     }
 
     @PutMapping
-    public ResponseEntity<?> putPost(@RequestBody Post post){
+    public ResponseEntity<?> putPost(@RequestBody Post post, @AuthenticationPrincipal User user){//글 수정
+
+        if(!user.equals(postService.getPost(post.getId()).getUser())){
+            return ResponseEntity.status(403).build();
+        }
+
         EntityModel<Post> body = EntityModel.of(postService.update(post));
 
-        body.add(linkTo(methodOn(PostController.class).putPost(post)).withSelfRel());
+        body.add(linkTo(methodOn(PostController.class).putPost(post, user)).withSelfRel());
         return ResponseEntity.ok(body);
     }
 
@@ -73,7 +79,12 @@ public class PostController {
     }
 
     @DeleteMapping("/{postId}")
-    public ResponseEntity<?> deletePost(@PathVariable("postId") Long postId){
+    public ResponseEntity<?> deletePost(@PathVariable("postId") Long postId, @AuthenticationPrincipal User user){
+
+        if(!user.equals(postService.getPost(postId).getUser()) || !user.getRoles().contains("ADMIN")){
+            return ResponseEntity.status(403).build();
+        }
+        
         return ResponseEntity.ok(postService.delete(postId));
     }
 }
